Menambahkan bid lelang

+++ database/migrations/2023_02_03_011346_create_history_lelangs_table.php
@@ -20,7 +20,7 @@ class CreateHistoryLelangsTable extends Migration
             $table->unsignedBigInteger('users_id');
             $table->foreign('users_id')->references('id')->on('users')->onDelete('cascade');
             $table->string('nama_barang');
-            $table->integer('harga');
+            $table->BigInteger('harga');
             $table->enum('status', ['pending', 'gugur', 'pemenang']);
             $table->timestamps();
         });
